Add forgotten files for introduction to new RM 4 Mini and Pro distinction.

+++ bundles/org.openhab.binding.broadlink/src/main/java/org/openhab/binding/broadlink/internal/BroadlinkProtocol.java
@@ -196,7 +196,7 @@ public class BroadlinkProtocol {
             throw new ProtocolException("Unexpectedly short packet; length " + packet.length
                     + " is shorter than protocol minimum " + MIN_RESPONSE_PACKET_LENGTH);
         }
-        boolean error = packet[0x22] != 0 || packet[0x23] != 0;
+        boolean error = (int) packet[0x22] != 0 || (int) packet[0x23] != 0;
         if (error) {
             throw new ProtocolException(String.format("Response from device is not valid. (0x22=0x%02X,0x23=0x%02X)",
                     packet[0x22], packet[0x23]));
